test: move PSM reviver next to PSM
